# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zniche API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type ProductStatus = (typeof ProductStatus)[keyof typeof ProductStatus];

export const ProductStatus = {
  building: "building",
  completed: "completed",
  failed: "failed",
} as const;

export interface Product {
  id: string;
  userId: string;
  skill: string;
  hoursPerWeek: number;
  price: number;
  status: ProductStatus;
  /** @nullable */
  productName?: string | null;
  /** @nullable */
  productDescription?: string | null;
  /** @nullable */
  productFormat?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  headline?: string | null;
  /** @nullable */
  salesCopy?: string | null;
  /** @nullable */
  socialCaptions?: string | null;
  /** @nullable */
  stripeCheckoutUrl?: string | null;
  /** @nullable */
  stripeProductId?: string | null;
  marketplaceListed?: boolean;
  /** @nullable */
  creatorFirstName?: string | null;
  /** @nullable */
  marketResearch?: string | null;
  createdAt: string;
}

export interface CreateProductBody {
  /** @minLength 1 */
  skill: string;
  /** @minimum 1 */
  hoursPerWeek: number;
  /**
   * @minimum 5
   * @maximum 500
   */
  price: number;
}

export interface UpdateProductBody {
  productName?: string;
  productDescription?: string;
  headline?: string;
  salesCopy?: string;
  stripeCheckoutUrl?: string;
  stripeProductId?: string;
  marketplaceListed?: boolean;
  status?: string;
  category?: string;
  productFormat?: string;
  socialCaptions?: string;
  marketResearch?: string;
}

export interface MarketplaceListing {
  id: string;
  productName: string;
  category: string;
  price: number;
  creatorFirstName: string;
  stripeCheckoutUrl: string;
  /** @nullable */
  productDescription?: string | null;
  /** @nullable */
  headline?: string | null;
  /** @nullable */
  productFormat?: string | null;
  createdAt: string;
}

export type MarketplaceStatsTopCategoriesItem = {
  category: string;
  count: number;
};

export interface MarketplaceStats {
  totalProducts: number;
  totalCreators: number;
  avgPrice: number;
  topCategories?: MarketplaceStatsTopCategoriesItem[];
}

export interface CreateCheckoutBody {
  productId: string;
}

export interface CheckoutResponse {
  checkoutUrl: string;
}

export interface AiBuildRequest {
  productId: string;
  skill: string;
  hoursPerWeek: number;
  price: number;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type GetMarketplaceListingsParams = {
  category?: string;
  sort?: GetMarketplaceListingsSort;
};

export type GetMarketplaceListingsSort =
  (typeof GetMarketplaceListingsSort)[keyof typeof GetMarketplaceListingsSort];

export const GetMarketplaceListingsSort = {
  newest: "newest",
  price_asc: "price_asc",
  price_desc: "price_desc",
} as const;