# lib/api-zod/src/generated/types/aiBuildRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zniche API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiBuildRequest {
  productId: string;
  skill: string;
  hoursPerWeek: number;
  price: number;
}