# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zniche API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  /** @minLength 1 */
  skill: string;
  /** @minimum 1 */
  hoursPerWeek: number;
  /**
   * @minimum 5
   * @maximum 500
   */
  price: number;
}