# lib/api-zod/src/generated/types/marketplaceListing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zniche API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MarketplaceListing {
  id: string;
  productName: string;
  category: string;
  price: number;
  creatorFirstName: string;
  stripeCheckoutUrl: string;
  /** @nullable */
  productDescription?: string | null;
  /** @nullable */
  headline?: string | null;
  /** @nullable */
  productFormat?: string | null;
  createdAt: string;
}