# lib/api-zod/src/generated/types/marketplaceStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zniche API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MarketplaceStatsTopCategoriesItem } from "./marketplaceStatsTopCategoriesItem";

export interface MarketplaceStats {
  totalProducts: number;
  totalCreators: number;
  avgPrice: number;
  topCategories?: MarketplaceStatsTopCategoriesItem[];
}