# lib/api-zod/src/generated/types/updateProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zniche API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProductBody {
  productName?: string;
  productDescription?: string;
  headline?: string;
  salesCopy?: string;
  stripeCheckoutUrl?: string;
  stripeProductId?: string;
  marketplaceListed?: boolean;
  status?: string;
  category?: string;
  productFormat?: string;
  socialCaptions?: string;
  marketResearch?: string;
}